requirements.txt
langchain
gradio
transformers
torch
python-dotenv
sentencepiece